feat: add cli link filters
