+++ package.json
@@ -22,6 +22,7 @@
     "async": "^3.2.5",
     "chalk": "^5.3.0",
     "cheerio": "^1.0.0-rc.12",
+    "commander": "^11.1.0",
     "dotenv": "^16.3.1",
     "moment": "^2.29.4",
     "next": "14.0.0",
